# docs/api/f026-openapi.yaml
openapi: 3.0.3
info:
  title: Theme Playground API
  description: |
    API for managing themes, accessibility validation, and user customization in the Theme Playground system.

    This API provides comprehensive theme management capabilities including:
    - Theme discovery and browsing
    - Live theme preview and application
    - Accessibility compliance validation
    - Custom theme creation and modification
    - Theme import/export functionality
    - User preference management

    The API is designed with accessibility-first principles and includes real-time WCAG compliance checking.
  version: 1.0.0
  contact:
    name: Theme System Team
    email: themes@go-redis-wq.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.go-redis-wq.com/v1
    description: Production server
  - url: https://staging-api.go-redis-wq.com/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
  /themes:
    get:
      summary: List available themes
      description: |
        Retrieve a paginated list of available themes with optional filtering by category,
        accessibility compliance, and search terms. Includes theme metadata and accessibility information.
      operationId: listThemes
      tags:
        - Themes
      parameters:
        - name: category
          in: query
          description: Filter themes by category
          required: false
          schema:
            $ref: '#/components/schemas/ThemeCategory'
        - name: accessibility
          in: query
          description: Filter by accessibility compliance level
          required: false
          schema:
            type: string
            enum: [A, AA, AAA]
        - name: search
          in: query
          description: Search themes by name, description, or tags
          required: false
          schema:
            type: string
            maxLength: 100
        - name: tags
          in: query
          description: Filter by specific tags (comma-separated)
          required: false
          schema:
            type: string
        - name: author
          in: query
          description: Filter themes by author
          required: false
          schema:
            type: string
            maxLength: 100
        - name: color_blind_safe
          in: query
          description: Filter themes that are color blind safe
          required: false
          schema:
            type: boolean
        - name: limit
          in: query
          description: Number of themes to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          description: Number of themes to skip for pagination
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: sort
          in: query
          description: Sort order for themes
          required: false
          schema:
            type: string
            enum: [name, popularity, created_at, updated_at, accessibility_score]
            default: name
        - name: sort_order
          in: query
          description: Sort direction
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Successfully retrieved theme list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThemeListResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create a new custom theme
      description: |
        Create a new custom theme based on an existing theme or from scratch.
        The theme will be validated for structure and accessibility compliance.
      operationId: createTheme
      tags:
        - Themes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateThemeRequest'
      responses:
        '201':
          description: Theme created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThemeResponse'
        '400':
          description: Invalid theme data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Theme name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Theme validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
  /themes/{themeId}:
    get:
      summary: Get theme details
      description: Retrieve complete theme definition including all style components and accessibility information
      operationId: getTheme
      tags:
        - Themes
      parameters:
        - name: themeId
          in: path
          description: Theme identifier
          required: true
          schema:
            type: string
            pattern: '^[a-z0-9-]+$'
        - name: include_accessibility
          in: query
          description: Include accessibility analysis in response
          required: false
          schema:
            type: boolean
            default: true
        - name: include_preview
          in: query
          description: Include preview data
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Theme details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThemeDetailResponse'
        '404':
          description: Theme not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update custom theme
      description: |
        Update an existing custom theme. Only the theme owner or administrators can modify themes.
        System themes cannot be modified.
      operationId: updateTheme
      tags:
        - Themes
      parameters:
        - name: themeId
          in: path
          description: Theme identifier
          required: true
          schema:
            type: string
            pattern: '^[a-z0-9-]+$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateThemeRequest'
      responses:
        '200':
          description: Theme updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThemeResponse'
        '400':
          description: Invalid theme data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '403':
          description: Not authorized to modify this theme
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Theme not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete custom theme
      description: |
        Delete a custom theme. Only the theme owner or administrators can delete themes.
        System themes cannot be deleted.
      operationId: deleteTheme
      tags:
        - Themes
      parameters:
        - name: themeId
          in: path
          description: Theme identifier
          required: true
          schema:
            type: string
            pattern: '^[a-z0-9-]+$'
      responses:
        '204':
          description: Theme deleted successfully
        '403':
          description: Not authorized to delete this theme
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Theme not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Cannot delete theme that is currently in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /themes/{themeId}/preview:
    post:
      summary: Preview theme without applying
      description: |
        Temporarily preview a theme without persisting the change. Returns rendered
        component examples and accessibility analysis.
      operationId: previewTheme
      tags:
        - Theme Application
      parameters:
        - name: themeId
          in: path
          description: Theme identifier
          required: true
          schema:
            type: string
            pattern: '^[a-z0-9-]+$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreviewThemeRequest'
      responses:
        '200':
          description: Theme preview generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThemePreviewResponse'
        '404':
          description: Theme not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /themes/{themeId}/export:
    get:
      summary: Export theme in various formats
      description: |
        Export a theme in different formats for use in other applications or for sharing.
        Supports JSON, YAML, CSS, and Lip Gloss formats.
      operationId: exportTheme
      tags:
        - Import/Export
      parameters:
        - name: themeId
          in: path
          description: Theme identifier
          required: true
          schema:
            type: string
            pattern: '^[a-z0-9-]+$'
        - name: format
          in: query
          description: Export format
          required: false
          schema:
            type: string
            enum: [json, yaml, css, lipgloss]
            default: json
        - name: include_metadata
          in: query
          description: Include theme metadata in export
          required: false
          schema:
            type: boolean
            default: true
        - name: minify
          in: query
          description: Minify exported content
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Theme exported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponse'
            application/yaml:
              schema:
                type: string
            text/css:
              schema:
                type: string
            text/plain:
              schema:
                type: string
        '404':
          description: Theme not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /themes/import:
    post:
      summary: Import theme from external source
      description: |
        Import a theme from a URL, file upload, or raw JSON data.
        The imported theme will be validated and can be saved as a custom theme.
      operationId: importTheme
      tags:
        - Import/Export
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportThemeRequest'
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Theme file to upload
                validate:
                  type: boolean
                  default: true
                save_as:
                  type: string
                  description: Name to save imported theme as
      responses:
        '201':
          description: Theme imported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportThemeResponse'
        '400':
          description: Invalid import data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '422':
          description: Theme validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Theme:
      type: object
      required:
        - metadata
        - palette
        - components
      properties:
        metadata:
          $ref: '#/components/schemas/ThemeMetadata'
        palette:
          $ref: '#/components/schemas/ColorPalette'
        components:
          $ref: '#/components/schemas/ComponentStyles'
        typography:
          $ref: '#/components/schemas/Typography'
        animations:
          $ref: '#/components/schemas/AnimationConfig'
        accessibility:
          $ref: '#/components/schemas/AccessibilityInfo'
...
    ThemeCategory:
      type: string
      enum:
        - standard
        - accessibility
        - specialty
        - custom
      description: Theme category classification
    ColorPalette:
      type: object
      required:
        - background
        - surface
        - primary
        - secondary
        - accent
        - success
        - warning
        - error
        - info
        - text_primary
        - text_secondary
        - border
      properties:
        # Base colors
        background:
          $ref: '#/components/schemas/Color'
        surface:
          $ref: '#/components/schemas/Color'
        surface_variant:
          $ref: '#/components/schemas/Color'
        primary:
          $ref: '#/components/schemas/Color'
        primary_variant:
          $ref: '#/components/schemas/Color'
        secondary:
          $ref: '#/components/schemas/Color'
        secondary_variant:
          $ref: '#/components/schemas/Color'
        accent:
          $ref: '#/components/schemas/Color'
        # Semantic colors
        success:
          $ref: '#/components/schemas/Color'
        warning:
          $ref: '#/components/schemas/Color'
        error:
          $ref: '#/components/schemas/Color'
        info:
          $ref: '#/components/schemas/Color'
        # Text hierarchy
        text_primary:
          $ref: '#/components/schemas/Color'
        text_secondary:
          $ref: '#/components/schemas/Color'
        text_tertiary:
          $ref: '#/components/schemas/Color'
        text_disabled:
          $ref: '#/components/schemas/Color'
        text_inverse:
          $ref: '#/components/schemas/Color'
        # Interactive elements
        border:
          $ref: '#/components/schemas/Color'
        border_focus:
          $ref: '#/components/schemas/Color'
        border_error:
          $ref: '#/components/schemas/Color'
        divider:
          $ref: '#/components/schemas/Color'
        shadow:
          $ref: '#/components/schemas/Color'
        overlay:
          $ref: '#/components/schemas/Color'
        # Status colors
        status_pending:
          $ref: '#/components/schemas/Color'
        status_running:
          $ref: '#/components/schemas/Color'
        status_completed:
          $ref: '#/components/schemas/Color'
        status_failed:
          $ref: '#/components/schemas/Color'
        status_retry:
          $ref: '#/components/schemas/Color'
...
    ComponentStyles:
      type: object
      properties:
        button:
          $ref: '#/components/schemas/ButtonStyles'
        input:
          $ref: '#/components/schemas/InputStyles'
        table:
          $ref: '#/components/schemas/TableStyles'
        card:
          $ref: '#/components/schemas/CardStyles'
        modal:
          $ref: '#/components/schemas/ModalStyles'
        navigation:
          $ref: '#/components/schemas/NavigationStyles'
        status:
          $ref: '#/components/schemas/StatusStyles'
        chart:
          $ref: '#/components/schemas/ChartStyles'
...
    Typography:
      type: object
      properties:
        font_family:
          type: string
          description: Primary font family
        font_size_base:
          type: integer
          minimum: 8
          maximum: 24
          description: Base font size in points
        line_height:
          type: number
          minimum: 1.0
          maximum: 3.0
          description: Line height multiplier
        letter_spacing:
          type: number
          description: Letter spacing adjustment
    AnimationConfig:
      type: object
      properties:
        enabled:
          type: boolean
          default: true
        duration_fast:
          type: integer
          minimum: 50
          maximum: 500
          description: Fast animation duration in milliseconds
        duration_normal:
          type: integer
          minimum: 100
          maximum: 1000
          description: Normal animation duration in milliseconds
        duration_slow:
          type: integer
          minimum: 200
          maximum: 2000
          description: Slow animation duration in milliseconds
        easing:
          type: string
          enum: [linear, ease, ease-in, ease-out, ease-in-out]
...
    ThemeListResponse:
      type: object
      required:
        - themes
        - total
        - categories
      properties:
        themes:
          type: array
          items:
            $ref: '#/components/schemas/ThemeSummary'
        total:
          type: integer
          minimum: 0
          description: Total number of themes matching filters
        categories:
          type: array
          items:
            $ref: '#/components/schemas/ThemeCategory'
        collections:
          type: array
          items:
            type: string
          description: Available theme collections
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    ThemeSummary:
      type: object
      required:
        - name
        - metadata
      properties:
        name:
          type: string
        metadata:
          $ref: '#/components/schemas/ThemeMetadata'
        accessibility:
          $ref: '#/components/schemas/AccessibilityInfo'
        preview_colors:
          type: array
          items:
            type: string
          maxItems: 5
          description: Sample colors for quick preview
        user_data:
          $ref: '#/components/schemas/UserThemeData'
    UserThemeData:
      type: object
      properties:
        is_favorite:
          type: boolean
        last_used:
          type: string
          format: date-time
        use_count:
          type: integer
          minimum: 0
    ThemeDetailResponse:
      type: object
      required:
        - theme
      properties:
        theme:
          $ref: '#/components/schemas/Theme'
        accessibility_report:
          $ref: '#/components/schemas/AccessibilityReport'
        preview_data:
          $ref: '#/components/schemas/ThemePreviewData'
        usage_analytics:
          $ref: '#/components/schemas/ThemeUsageAnalytics'
    ThemeResponse:
      type: object
      required:
        - theme
      properties:
        theme:
          $ref: '#/components/schemas/Theme'
        validation_warnings:
          type: array
          items:
            type: string
    CreateThemeRequest:
      type: object
      required:
        - name
        - palette
      properties:
        name:
          type: string
          pattern: '^[a-z0-9-]+$'
          minLength: 1
          maxLength: 50
        description:
          type: string
          maxLength: 200
        based_on:
          type: string
          description: Existing theme to use as base
        palette:
          $ref: '#/components/schemas/ColorPalette'
        components:
          $ref: '#/components/schemas/ComponentStyles'
        typography:
          $ref: '#/components/schemas/Typography'
        animations:
          $ref: '#/components/schemas/AnimationConfig'
        tags:
          type: array
          items:
            type: string
            pattern: '^[a-z0-9-]+$'
          maxItems: 10
        validate_accessibility:
          type: boolean
          default: true
    UpdateThemeRequest:
      type: object
      properties:
        description:
          type: string
          maxLength: 200
        palette:
          $ref: '#/components/schemas/ColorPalette'
        components:
          $ref: '#/components/schemas/ComponentStyles'
        typography:
          $ref: '#/components/schemas/Typography'
        animations:
          $ref: '#/components/schemas/AnimationConfig'
        tags:
          type: array
          items:
            type: string
            pattern: '^[a-z0-9-]+$'
          maxItems: 10
...
    PreviewThemeRequest:
      type: object
      properties:
        components:
          type: array
          items:
            type: string
          description: Specific components to include in preview
        include_accessibility:
          type: boolean
          default: true
    ThemePreviewResponse:
      type: object
      required:
        - theme
        - preview_data
      properties:
        theme:
          type: string
        preview_data:
          $ref: '#/components/schemas/ThemePreviewData'
        accessibility_report:
          $ref: '#/components/schemas/AccessibilityReport'
    ThemePreviewData:
      type: object
      properties:
        components:
          type: object
          additionalProperties:
            type: string
          description: Rendered component examples
        color_swatches:
          type: array
          items:
            $ref: '#/components/schemas/ColorSwatch'
        typography_samples:
          type: object
          additionalProperties:
            type: string
...
    ImportThemeRequest:
      oneOf:
        - type: object
          required:
            - source
            - url
          properties:
            source:
              type: string
              enum: [url]
            url:
              type: string
              format: uri
        - type: object
          required:
            - source
            - data
          properties:
            source:
              type: string
              enum: [json]
            data:
              type: object
      properties:
        validate:
          type: boolean
          default: true
        save_as:
          type: string
          pattern: '^[a-z0-9-]+$'
          description: Name to save imported theme as
    ImportThemeResponse:
      type: object
      required:
        - imported
        - theme
      properties:
        imported:
          type: boolean
        theme:
          $ref: '#/components/schemas/Theme'
        validation_report:
          $ref: '#/components/schemas/ValidationReport'
        accessibility_report:
          $ref: '#/components/schemas/AccessibilityReport'
    ExportResponse:
      type: object
      required:
        - format
        - data
      properties:
        format:
          type: string
          enum: [json, yaml, css, lipgloss]
        data:
          type: string
        metadata:
          type: object
          properties:
            theme_name:
              type: string
            exported_at:
              type: string
              format: date-time
            version:
              type: string
...
    PaginationInfo:
      type: object
      required:
        - limit
        - offset
        - total
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
        has_next:
          type: boolean
        has_previous:
          type: boolean
...
    ValidationError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string
        code:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        request_id:
          type: string
    ValidationErrorResponse:
      type: object
      required:
        - error
        - message
        - validation_errors
      properties:
        error:
          type: string
        message:
          type: string
        validation_errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        request_id:
          type: string
...